consolidates content component
